plus with long instead of int
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@16730 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 2152bf66f1fe442b66e33e6831c2d2c918ff8059

+++ core/core/src/de/uniol/inf/is/odysseus/core/metadata/PointInTime.java
@@ -265,7 +265,7 @@ public class PointInTime implements Comparable<PointInTime>, Cloneable,
 		return new PointInTime(time.point + toAdd);
 	}
 
-	public PointInTime plus(int time) {
+	public PointInTime plus(long time) {
 		if (isInfinite)
 			return this;
 		return new PointInTime(this.point + time);
